safari-600.7-branch/LayoutTests/http/tests/security/script-crossorigin-onerror-information.html

<!DOCTYPE html>
<head>
</head>
<body>
<script src="../../js-test-resources/js-test-pre.js"></script>
<script>
window.jsTestIsAsync = true;
description("The test passes if window.onerror gets unsanitized information about the script error.");

window.onerror = function(msg, url, line) {
    window.msg = msg;
    window.url = url;
    window.line = line;
    shouldBeEqualToString("msg.match(/SomeError/)[0]", "SomeError");
    shouldBeEqualToString("url", "http://localhost:8000/security/resources/cors-script.php?fail=true");
    shouldBe("line", "1");
    finishJSTest();
}
</script>
<script crossorigin="    anonymous " src="http://localhost:8000/security/resources/cors-script.php?fail=true"></script>
<script src="../../js-test-resources/js-test-post.js"></script>
